requirements.txt
pytest-benchmark
ujson==1.35
simplejson==3.10.0
python-rapidjson==0.0.6
msgpack-python==0.4.8
invoke==0.13.0
pyrobuf==0.5.11